Remove obsolete classes/methods
