+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/classfile/ClassfileParser.java
@@ -20,6 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
+
 package com.oracle.truffle.espresso.classfile;
 
 import static com.oracle.truffle.espresso.EspressoOptions.SpecComplianceMode.STRICT;
@@ -55,7 +56,7 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
-import com.oracle.truffle.espresso.verifier.MethodVerifier;
+import com.oracle.truffle.espresso.impl.ClassLoadingEnv;
 import org.graalvm.collections.EconomicMap;
 
 import com.oracle.truffle.espresso.classfile.ConstantPool.Tag;
@@ -99,6 +100,7 @@ import com.oracle.truffle.espresso.runtime.Attribute;
 import com.oracle.truffle.espresso.runtime.ClasspathFile;
 import com.oracle.truffle.espresso.runtime.EspressoContext;
 import com.oracle.truffle.espresso.runtime.EspressoException;
+import com.oracle.truffle.espresso.runtime.JavaVersion;
 import com.oracle.truffle.espresso.runtime.StaticObject;
 
 @SuppressWarnings("try")
@@ -160,7 +162,7 @@ public final class ClassfileParser {
 
     private final Symbol<Type> requestedClassType;
 
-    private final EspressoContext context;
+    private final ClassLoadingEnv env;
 
     private final ClassfileStream stream;
 
@@ -179,9 +181,9 @@ public final class ClassfileParser {
 
     private ConstantPool pool;
 
-    private ClassfileParser(ClassfileStream stream, StaticObject loader, Symbol<Type> requestedClassType, EspressoContext context, ClassRegistry.ClassDefinitionInfo info) {
+    private ClassfileParser(ClassLoadingEnv env, ClassfileStream stream, StaticObject loader, Symbol<Type> requestedClassType, ClassRegistry.ClassDefinitionInfo info) {
+        this.env = env;
         this.requestedClassType = requestedClassType;
-        this.context = context;
         this.classfile = null;
         this.stream = Objects.requireNonNull(stream);
         this.loader = loader;
@@ -189,8 +191,8 @@ public final class ClassfileParser {
     }
 
     // Note: only used for reading the class name from class bytes
-    private ClassfileParser(ClassfileStream stream, EspressoContext context) {
-        this(stream, null, null, context, ClassRegistry.ClassDefinitionInfo.EMPTY);
+    private ClassfileParser(ClassLoadingEnv env, ClassfileStream stream) {
+        this(env, stream, null, null, ClassRegistry.ClassDefinitionInfo.EMPTY);
     }
 
     void handleBadConstant(Tag tag, ClassfileStream s) {
@@ -222,21 +224,21 @@ public final class ClassfileParser {
         }
     }
 
-    public static ParserKlass parse(ClassfileStream stream, StaticObject loader, Symbol<Type> requestedClassName, EspressoContext context) {
-        return parse(stream, loader, requestedClassName, context, ClassRegistry.ClassDefinitionInfo.EMPTY);
+    public static ParserKlass parse(ClassLoadingEnv env, ClassfileStream stream, StaticObject loader, Symbol<Type> requestedClassName) {
+        return parse(env, stream, loader, requestedClassName, ClassRegistry.ClassDefinitionInfo.EMPTY);
     }
 
-    public static ParserKlass parse(ClassfileStream stream, StaticObject loader, Symbol<Type> requestedClassType, EspressoContext context, ClassRegistry.ClassDefinitionInfo info) {
-        return new ClassfileParser(stream, loader, requestedClassType, context, info).parseClass();
+    public static ParserKlass parse(ClassLoadingEnv env, ClassfileStream stream, StaticObject loader, Symbol<Type> requestedClassType, ClassRegistry.ClassDefinitionInfo info) {
+        return new ClassfileParser(env, stream, loader, requestedClassType, info).parseClass();
     }
 
     private ParserKlass parseClass() {
-        try (DebugCloseable parse = KLASS_PARSE.scope(context.getTimers())) {
+        try (DebugCloseable parse = KLASS_PARSE.scope(env.getTimers())) {
             return parseClassImpl();
         } catch (EspressoException e) {
             throw e;
         } catch (Throwable e) {
-            context.getLogger().severe("Unexpected host exception " + e + " thrown during class parsing.");
+            env.getLogger().severe("Unexpected host exception " + e + " thrown during class parsing.");
             throw e;
         }
     }
@@ -255,12 +257,13 @@ public final class ClassfileParser {
      * @param minor the minor version number
      */
     private void verifyVersion(int major, int minor) {
-        if (context.getJavaVersion().java8OrEarlier()) {
-            versionCheck8OrEarlier(context.getJavaVersion().classFileVersion(), major, minor);
-        } else if (context.getJavaVersion().java11OrEarlier()) {
-            versionCheck11OrEarlier(context.getJavaVersion().classFileVersion(), major, minor);
+        JavaVersion version = env.getJavaVersion();
+        if (version.java8OrEarlier()) {
+            versionCheck8OrEarlier(version.classFileVersion(), major, minor);
+        } else if (version.java11OrEarlier()) {
+            versionCheck11OrEarlier(version.classFileVersion(), major, minor);
         } else {
-            versionCheck12OrLater(context.getJavaVersion().classFileVersion(), major, minor);
+            versionCheck12OrLater(version.classFileVersion(), major, minor);
         }
     }
 
@@ -345,9 +348,9 @@ public final class ClassfileParser {
         final boolean majorGte15 = majorVersion >= JAVA_1_5_VERSION;
 
         if ((isAbstract && isFinal) ||
-                        (isInterface && !isAbstract) ||
-                        (isInterface && majorGte15 && (isSuper || isEnum)) ||
-                        (!isInterface && majorGte15 && isAnnotation)) {
+                (isInterface && !isAbstract) ||
+                (isInterface && majorGte15 && (isSuper || isEnum)) ||
+                (!isInterface && majorGte15 && isAnnotation)) {
             throw ConstantPool.classFormatError("Invalid class flags 0x" + Integer.toHexString(flags));
         }
     }
@@ -359,8 +362,8 @@ public final class ClassfileParser {
         majorVersion = stream.readU2();
         verifyVersion(majorVersion, minorVersion);
 
-        try (DebugCloseable closeable = CONSTANT_POOL.scope(context.getTimers())) {
-            this.pool = ConstantPool.parse(context.getLanguage(), stream, this, classDefinitionInfo.patches, context, majorVersion, minorVersion);
+        try (DebugCloseable closeable = CONSTANT_POOL.scope(env.getTimers())) {
+            this.pool = ConstantPool.parse(env, stream, this, classDefinitionInfo.patches, majorVersion, minorVersion);
         }
 
         // JVM_ACC_MODULE is defined in JDK-9 and later.
@@ -398,7 +401,7 @@ public final class ClassfileParser {
 
         // TODO(peterssen): Verify class names.
 
-        Symbol<Type> thisKlassType = context.getTypes().fromName(thisKlassName);
+        Symbol<Type> thisKlassType = env.getTypes().fromName(thisKlassName);
         if (Types.isPrimitive(thisKlassType) || Types.isArray(thisKlassType)) {
             throw ConstantPool.classFormatError(".this_class cannot be array nor primitive " + classType);
         }
@@ -422,19 +425,19 @@ public final class ClassfileParser {
         }
 
         Symbol<Type>[] superInterfaces;
-        try (DebugCloseable closeable = PARSE_INTERFACES.scope(context.getTimers())) {
+        try (DebugCloseable closeable = PARSE_INTERFACES.scope(env.getTimers())) {
             superInterfaces = parseInterfaces();
         }
         final ParserField[] fields;
-        try (DebugCloseable closeable = PARSE_FIELD.scope(context.getTimers())) {
+        try (DebugCloseable closeable = PARSE_FIELD.scope(env.getTimers())) {
             fields = parseFields(isInterface);
         }
         final ParserMethod[] methods;
-        try (DebugCloseable closeable = PARSE_METHODS.scope(context.getTimers())) {
+        try (DebugCloseable closeable = PARSE_METHODS.scope(env.getTimers())) {
             methods = parseMethods(isInterface);
         }
         final Attribute[] attributes;
-        try (DebugCloseable closeable = PARSE_CLASSATTR.scope(context.getTimers())) {
+        try (DebugCloseable closeable = PARSE_CLASSATTR.scope(env.getTimers())) {
             attributes = parseClassAttributes();
         }
 
@@ -443,18 +446,18 @@ public final class ClassfileParser {
 
         if (classDefinitionInfo.isHidden()) {
             assert requestedClassType != null;
-            int futureKlassID = context.getLanguage().getNewKlassId();
+            int futureKlassID = env.getLanguage().getNewKlassId();
             classDefinitionInfo.initKlassID(futureKlassID);
-            thisKlassName = context.getNames().getOrCreate(Types.hiddenClassName(requestedClassType, futureKlassID));
-            thisKlassType = context.getTypes().fromName(thisKlassName);
+            thisKlassName = env.getNames().getOrCreate(Types.hiddenClassName(requestedClassType, futureKlassID));
+            thisKlassType = env.getTypes().fromName(thisKlassName);
             pool = pool.patchForHiddenClass(thisKlassIndex, thisKlassName);
         }
 
         return new ParserKlass(pool, classDefinitionInfo.patchFlags(classFlags), thisKlassName, thisKlassType, superKlass, superInterfaces, methods, fields, attributes, thisKlassIndex);
     }
 
-    public static Symbol<Symbol.Name> getClassName(byte[] bytes, EspressoContext context) {
-        return new ClassfileParser(new ClassfileStream(bytes, null), context).getClassName();
+    public static Symbol<Symbol.Name> getClassName(ClassLoadingEnv env, byte[] bytes) {
+        return new ClassfileParser(env, new ClassfileStream(bytes, null)).getClassName();
     }
 
     private Symbol<Symbol.Name> getClassName() {
@@ -465,8 +468,8 @@ public final class ClassfileParser {
         majorVersion = stream.readU2();
         verifyVersion(majorVersion, minorVersion);
 
-        try (DebugCloseable closeable = CONSTANT_POOL.scope(context.getTimers())) {
-            this.pool = ConstantPool.parse(context.getLanguage(), stream, this, classDefinitionInfo.patches, context, majorVersion, minorVersion);
+        try (DebugCloseable closeable = CONSTANT_POOL.scope(env.getTimers())) {
+            this.pool = ConstantPool.parse(env, stream, this, classDefinitionInfo.patches, majorVersion, minorVersion);
         }
 
         // JVM_ACC_MODULE is defined in JDK-9 and later.
@@ -494,11 +497,11 @@ public final class ClassfileParser {
         final HashSet<MethodKey> dup = new HashSet<>(methodCount);
         for (int i = 0; i < methodCount; ++i) {
             ParserMethod method;
-            try (DebugCloseable closeable = PARSE_SINGLE_METHOD.scope(context.getTimers())) {
+            try (DebugCloseable closeable = PARSE_SINGLE_METHOD.scope(env.getTimers())) {
                 method = parseMethod(isInterface);
             }
             methods[i] = method;
-            try (DebugCloseable closeable = NO_DUP_CHECK.scope(context.getTimers())) {
+            try (DebugCloseable closeable = NO_DUP_CHECK.scope(env.getTimers())) {
                 if (!dup.add(new MethodKey(method))) {
                     throw ConstantPool.classFormatError("Duplicate method name and signature: " + method.getName() + " " + method.getSignature());
                 }
@@ -600,11 +603,11 @@ public final class ClassfileParser {
 
     private static final int classAnnotations = RUNTIME_VISIBLE_ANNOTATIONS | RUNTIME_INVISIBLE_ANNOTATIONS | RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | SIGNATURE;
     private static final int methodAnnotations = RUNTIME_VISIBLE_ANNOTATIONS | RUNTIME_INVISIBLE_ANNOTATIONS | RUNTIME_VISIBLE_PARAMETER_ANNOTATIONS | RUNTIME_INVISIBLE_PARAMETER_ANNOTATIONS |
-                    RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | ANNOTATION_DEFAULT | SIGNATURE;
+            RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | ANNOTATION_DEFAULT | SIGNATURE;
     private static final int fieldAnnotations = RUNTIME_VISIBLE_ANNOTATIONS | RUNTIME_INVISIBLE_ANNOTATIONS | RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | SIGNATURE;
     private static final int codeAnnotations = RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | SIGNATURE;
     private static final int recordAnnotations = RUNTIME_VISIBLE_ANNOTATIONS | RUNTIME_INVISIBLE_ANNOTATIONS |
-                    RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | SIGNATURE;
+            RUNTIME_VISIBLE_TYPE_ANNOTATIONS | RUNTIME_INVISIBLE_TYPE_ANNOTATIONS | SIGNATURE;
 
     public enum InfoType {
         Class(classAnnotations),
@@ -678,7 +681,7 @@ public final class ClassfileParser {
                 }
                 return annotationDefault = new Attribute(attributeName, stream.readByteArray(attributeSize));
             } else if (infoType.supports(SIGNATURE) && attributeName.equals(Name.Signature)) {
-                if (context.getJavaVersion().java9OrLater() && signature != null) {
+                if (env.getJavaVersion().java9OrLater() && signature != null) {
                     throw ConstantPool.classFormatError("Duplicate AnnotationDefault attribute");
                 }
                 if (attributeSize != 2) {
@@ -703,9 +706,9 @@ public final class ClassfileParser {
         boolean isClinit = false;
         boolean isInit = false;
 
-        try (DebugCloseable closeable = METHOD_INIT.scope(context.getTimers())) {
+        try (DebugCloseable closeable = METHOD_INIT.scope(env.getTimers())) {
 
-            try (DebugCloseable nameCheck = NAME_CHECK.scope(context.getTimers())) {
+            try (DebugCloseable nameCheck = NAME_CHECK.scope(env.getTimers())) {
                 pool.utf8At(nameIndex).validateMethodName(true);
                 name = pool.symbolAt(nameIndex, "method name");
 
@@ -719,7 +722,7 @@ public final class ClassfileParser {
                         methodFlags = ACC_STATIC;
                     } else if ((methodFlags & ACC_STATIC) == ACC_STATIC) {
                         methodFlags &= (ACC_STRICT | ACC_STATIC);
-                    } else if (context.getJavaVersion().java9OrLater()) {
+                    } else if (env.getJavaVersion().java9OrLater()) {
                         throw ConstantPool.classFormatError("Method <clinit> is not static.");
                     }
                     // extraFlags = INITIALIZER | methodFlags;
@@ -747,7 +750,7 @@ public final class ClassfileParser {
              * In a class file whose version number is 51.0 or above, the method has its ACC_STATIC
              * flag set and takes no arguments (4.6).
              */
-            try (DebugCloseable signatureCheck = SIGNATURE_CHECK.scope(context.getTimers())) {
+            try (DebugCloseable signatureCheck = SIGNATURE_CHECK.scope(env.getTimers())) {
                 // Checks for void method if init or clinit.
                 /*
                  * Obtain slot number for the signature. Forces a validation, but better in startup
@@ -794,7 +797,7 @@ public final class ClassfileParser {
                 if (codeAttribute != null) {
                     throw ConstantPool.classFormatError("Duplicate Code attribute");
                 }
-                try (DebugCloseable code = CODE_PARSE.scope(context.getTimers())) {
+                try (DebugCloseable code = CODE_PARSE.scope(env.getTimers())) {
                     methodAttributes[i] = codeAttribute = parseCodeAttribute(attributeName);
                 }
             } else if (attributeName.equals(Name.Exceptions)) {
@@ -822,10 +825,10 @@ public final class ClassfileParser {
                         if (Type.java_lang_invoke_LambdaForm$Compiled.equals(annotType)) {
                             methodFlags |= ACC_LAMBDA_FORM_COMPILED;
                         } else if (Type.java_lang_invoke_LambdaForm$Hidden.equals(annotType) ||
-                                        Type.jdk_internal_vm_annotation_Hidden.equals(annotType)) {
+                                Type.jdk_internal_vm_annotation_Hidden.equals(annotType)) {
                             methodFlags |= ACC_HIDDEN;
                         } else if (Type.sun_reflect_CallerSensitive.equals(annotType) ||
-                                        Type.jdk_internal_reflect_CallerSensitive.equals(annotType)) {
+                                Type.jdk_internal_reflect_CallerSensitive.equals(annotType)) {
                             methodFlags |= ACC_CALLER_SENSITIVE;
                         } else if (Type.java_lang_invoke_ForceInline.equals(annotType) ||
                                         Type.jdk_internal_vm_annotation_ForceInline.equals(annotType)) {
@@ -1214,7 +1217,7 @@ public final class ClassfileParser {
              * HotSpot does not perform this check. Enforcing the spec here break some applications
              * in the wild e.g. Intellij IDEA.
              */
-            if (context.getLanguage().getSpecComplianceMode() == STRICT) {
+            if (env.getSpecComplianceMode() == STRICT) {
                 if (majorVersion >= JAVA_7_VERSION && innerClassInfo.innerNameIndex == 0 && outerClassIndex != 0) {
                     throw ConstantPool.classFormatError("InnerClassesAttribute: the value of the outer_class_info_index item must be zero if the value of the inner_name_index item is zero.");
                 }
@@ -1237,11 +1240,11 @@ public final class ClassfileParser {
                         throw ConstantPool.classFormatError("Duplicate entry in InnerClasses attribute");
                     }
                     if (innerClassIndex == otherInnerClassInfo.innerClassIndex ||
-                                    // The same class can be referenced by two different CP indices,
-                                    // compare by name instead.
-                                    (innerClassIndex != 0 &&
-                                                    otherInnerClassInfo.innerClassIndex != 0 &&
-                                                    innerClassName.equals(pool.classAt(otherInnerClassInfo.innerClassIndex).getName(pool)))) {
+                            // The same class can be referenced by two different CP indices,
+                            // compare by name instead.
+                            (innerClassIndex != 0 &&
+                                    otherInnerClassInfo.innerClassIndex != 0 &&
+                                    innerClassName.equals(pool.classAt(otherInnerClassInfo.innerClassIndex).getName(pool)))) {
                         duplicateInnerClass = true;
                     }
                 }
@@ -1251,9 +1254,9 @@ public final class ClassfileParser {
         if (duplicateInnerClass || hasCycles(innerClassInfos)) {
             // Mimic HotSpot: Ignore the whole InnerClasses attribute and return an empty one.
             final String cause = duplicateInnerClass
-                            ? "Duplicate inner_class_info_index (class names)"
-                            : "Cycle detected";
-            context.getLogger().warning(cause + " in InnerClassesAttribute, in class " + classType);
+                    ? "Duplicate inner_class_info_index (class names)"
+                    : "Cycle detected";
+            env.getLogger().warning(cause + " in InnerClassesAttribute, in class " + classType);
             return new InnerClassesAttribute(name, new InnerClassesAttribute.Entry[0]);
         }
 
@@ -1306,7 +1309,7 @@ public final class ClassfileParser {
     }
 
     private StackMapTableAttribute parseStackMapTableAttribute(Symbol<Name> attributeName, int attributeSize) {
-        if (MethodVerifier.needsVerify(context.getLanguage(), loader)) {
+        if (env.needsVerify(loader)) {
             return new StackMapTableAttribute(attributeName, stream.readByteArray(attributeSize));
         }
         stream.skip(attributeSize);
@@ -1388,7 +1391,7 @@ public final class ClassfileParser {
             innerClassAccessFlags |= ACC_ABSTRACT;
         }
 
-        if (innerClassIndex != 0 || context.getJavaVersion().java9OrLater()) {
+        if (innerClassIndex != 0 || env.getJavaVersion().java9OrLater()) {
             pool.classAt(innerClassIndex).validate(pool);
         }
         if (outerClassIndex != 0) {
@@ -1420,11 +1423,11 @@ public final class ClassfileParser {
         }
 
         byte[] code;
-        try (DebugCloseable codeRead = CODE_READ.scope(context.getTimers())) {
+        try (DebugCloseable codeRead = CODE_READ.scope(env.getTimers())) {
             code = stream.readByteArray(codeLength);
         }
         ExceptionHandler[] entries;
-        try (DebugCloseable handlers = EXCEPTION_HANDLERS.scope(context.getTimers())) {
+        try (DebugCloseable handlers = EXCEPTION_HANDLERS.scope(env.getTimers())) {
             entries = parseExceptionHandlerEntries();
         }
 
@@ -1517,7 +1520,7 @@ public final class ClassfileParser {
             int catchTypeIndex = stream.readU2();
             Symbol<Type> catchType = null;
             if (catchTypeIndex != 0) {
-                catchType = context.getTypes().fromName(pool.classAt(catchTypeIndex).getName(pool));
+                catchType = env.getTypes().fromName(pool.classAt(catchTypeIndex).getName(pool));
             }
             entries[i] = new ExceptionHandler(startPc, endPc, handlerPc, catchTypeIndex, catchType);
         }
@@ -1684,7 +1687,7 @@ public final class ClassfileParser {
             }
             return null;
         }
-        return context.getTypes().fromName(pool.classAt(index).getName(pool));
+        return env.getTypes().fromName(pool.classAt(index).getName(pool));
     }
 
     @SuppressWarnings({"unchecked", "rawtypes"})
@@ -1697,7 +1700,7 @@ public final class ClassfileParser {
         for (int i = 0; i < interfaceCount; i++) {
             int interfaceIndex = stream.readU2();
             Symbol<Name> interfaceName = pool.classAt(interfaceIndex).getName(pool);
-            Symbol<Type> interfaceType = context.getTypes().fromName(interfaceName);
+            Symbol<Type> interfaceType = env.getTypes().fromName(interfaceName);
             if (interfaceType == null) {
                 throw ConstantPool.classFormatError(classType + " contains invalid superinterface name: " + interfaceName);
             }
