+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/impl/LinkedKlass.java
@@ -99,20 +99,6 @@ public final class LinkedKlass {
         this.methods = linkedMethods;
     }
 
-    // TODO (ivan-ristovic): Remove
-    public static LinkedKlass create(ContextDescription description, ParserKlass parserKlass, LinkedKlass superKlass, LinkedKlass[] interfaces) {
-        LinkedKlassFieldLayout fieldLayout = new LinkedKlassFieldLayout(description, parserKlass, superKlass);
-        return new LinkedKlass(
-                parserKlass,
-                superKlass,
-                interfaces,
-                fieldLayout.instanceShape,
-                fieldLayout.staticShape,
-                fieldLayout.instanceFields,
-                fieldLayout.staticFields,
-                fieldLayout.fieldTableLength);
-    }
-
     public static LinkedKlass create(ClassLoadingEnv env, ParserKlass parserKlass, LinkedKlass superKlass, LinkedKlass[] interfaces) {
         LinkedKlassFieldLayout fieldLayout = new LinkedKlassFieldLayout(env, parserKlass, superKlass);
         return new LinkedKlass(
